requirements.txt
<<<<<<< HEAD
streamlit
google-generativeai
requests
=======
streamlit>=1.40.0
google-generativeai>=0.10.0
requests
>>>>>>> 05b497a (Fix: Add missing imports for requests and json, and update requirements.txt)
